+++ src/actuator/data/WindowBlind.java
@@ -2,12 +2,18 @@ package actuator.data;
 
 import actuator.common.Action;
 
+/**
+ * Acción que representa el ajuste de la persiana.
+ * El índice es abstracto, por lo que el dispositivo cliente
+ * los puede adaptar a la máquina que realmente controle.
+ */
 public final class WindowBlind extends Action {
 	
 	private String id;
-	private int angle;
+	// El prototipo modelará esto con un servo.
+	private byte angle;
 	
-	public WindowBlind(String id, int angle) {
+	public WindowBlind(String id, byte angle) {
 		super();
 		this.id = id;
 		this.angle = angle;
@@ -17,7 +23,7 @@ public final class WindowBlind extends Action {
 		return id;
 	}
 
-	public int angle() {
+	public byte angle() {
 		return angle;
 	}
